requirements.txt
astroid==2.2.5
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.1
dj-database-url==0.5.0
Django==2.2.4
django-cors-headers==3.1.0
django-heroku==0.3.1
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
djwebsockets==0.9.6
gunicorn==19.9.0
idna==2.8
isort==4.3.21
joblib==0.13.2
lazy-object-proxy==1.4.1
mccabe==0.6.1
numpy==1.17.1
pandas==0.25.1
Pillow==6.1.0
psycopg2==2.8.3
PyJWT==1.7.1
pylint==2.3.1
python-dateutil==2.8.0
pytz==2019.2
requests==2.22.0
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
sklearn==0.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.3
whitenoise==4.1.3
wrapt==1.11.2
